Ajuste de variables
Ahora las variables usadas son las correspondientes, nada de utilizar
int para valores pequeños.

+++ TestGraph/colisiones.c
@@ -1,6 +1,6 @@
 #include "cabecera.h"
 
-int colisiones(char comp)
+short colisiones(char comp)
 {
 	if (comp == '#')
 	{
